src/orderservice/requirements.txt
cherrypy
tomlkit
cherrypy_cors
requests
urllib3
psycopg2
stripe
sqlalchemy
SQLAlchemy-serializer
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-requests
